refactored code to include try-catch exceptions

+++ src/main/java/org/personal/SimpleDBViewer/CRUDRepository/CPUListEntityCRUDRepository.java
@@ -7,11 +7,21 @@ import org.hibernate.SessionFactory;
 import org.personal.SimpleDBViewer.domain.CPUListEntity;
 
 public class CPUListEntityCRUDRepository {
-	public static void createCPUListEntity(SessionFactory s, CPUListEntity cpu) {
+	public static void createCPUListEntity(SessionFactory s, CPUListEntity cpu) throws IllegalArgumentException {
+		// validate input
+		if(cpu == null) {
+			throw new IllegalArgumentException("Input CPUListEntity object cannot be null");
+		}
+
 		AbstractCRUDRepository.createEntity(s, cpu);
 	}
 
-	public static CPUListEntity createCPUListEntity(SessionFactory s, String cpuName) {
+	public static CPUListEntity createCPUListEntity(SessionFactory s, String cpuName) throws IllegalArgumentException {
+		// validate input
+		if(cpuName.isEmpty()) {
+			throw new IllegalArgumentException("Input cpuName cannot be empty/null.");
+		}
+
 		CPUListEntity cpu = new CPUListEntity();
 		cpu.setName(cpuName);
 		AbstractCRUDRepository.createEntity(s, cpu);
@@ -19,10 +29,15 @@ public class CPUListEntityCRUDRepository {
 	}
 
 	public static CPUListEntity getCPU(SessionFactory s, CPUListEntity cpu) {
-		return getCPU(s, cpu.getName());
+		return getCPU(s, cpu.getId());
 	}
 
-	public static CPUListEntity getCPU(SessionFactory s, Long cpuId) {
+	public static CPUListEntity getCPU(SessionFactory s, Long cpuId) throws IllegalArgumentException {
+		// validate input
+		if(cpuId == null) {
+			throw new IllegalArgumentException("Input cpuId cannot be empty/null.");
+		}
+
 		Session session = s.openSession();
 		CPUListEntity rtnCPU = null;
 		try {
@@ -40,7 +55,12 @@ public class CPUListEntityCRUDRepository {
 		return rtnCPU;
 	}
 
-	public static CPUListEntity getCPU(SessionFactory s, String cpuName) {
+	public static CPUListEntity getCPU(SessionFactory s, String cpuName) throws IllegalArgumentException {
+		// validate input
+		if(cpuName.isEmpty()) {
+			throw new IllegalArgumentException("Input cpuName cannot be empty/null");
+		}
+
 		Session session = s.openSession();
 		CPUListEntity rtnCPU = null;
 		try {
@@ -61,18 +81,34 @@ public class CPUListEntityCRUDRepository {
 	public static List<CPUListEntity> getAllCPUs(SessionFactory s) {
 		Session session = s.openSession();
 		List<CPUListEntity> rtnList = null;
-		session.beginTransaction();
-		rtnList = session.createSelectionQuery("from CPUListEntity", CPUListEntity.class).getResultList();
-		session.getTransaction().commit();
-		session.close();
+		try {
+			session.beginTransaction();
+			rtnList = session.createSelectionQuery("from CPUListEntity", CPUListEntity.class).getResultList();
+			session.getTransaction().commit();
+		} catch (Exception e) {
+			if(session.getTransaction().isActive()) session.getTransaction().rollback();
+			throw e;
+		} finally {
+			session.close();
+		}
 		return rtnList;
 	}
 
-	public static void updateCPUListEntity(SessionFactory s, CPUListEntity cpu) {
+	public static void updateCPUListEntity(SessionFactory s, CPUListEntity cpu) throws IllegalArgumentException {
+		// validate input
+		if(cpu == null) {
+			throw new IllegalArgumentException("Input CPUListEntity object cannot be null");
+		}
+
 		AbstractCRUDRepository.updateEntity(s, cpu);
 	}
 
-	public static void deleteCPUListEntity(SessionFactory s, CPUListEntity cpu) {
+	public static void deleteCPUListEntity(SessionFactory s, CPUListEntity cpu) throws IllegalArgumentException {
+		// validate input
+		if(cpu == null) {
+			throw new IllegalArgumentException("Input CPUListEntity object cannot be null");
+		}
+
 		AbstractCRUDRepository.deleteEntity(s, cpu);
 	}
 }
